# apps/docs/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Tributary API
  version: 1.9.0
  description: >-
    Modular Express API for subscription and payment services on Solana.
    Provides health checks, subscription status lookups, on-chain event queries,
    webhook management, JWT issuance, JWKS publishing, and admin key rotation.
  license:
    name: MIT
servers:
  - url: https://api.tributary.so
    description: Production
  - url: http://localhost:3002
    description: Local development
components:
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: x-admin-key
      description: Admin API key (ADMIN_API_KEY env var on the server).
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        timestamp:
          type: integer
          description: Unix epoch milliseconds.
...
paths:
...
  /v1/onetime/{trackingId}:
    get:
      summary: Look up one-time payment
      description: >
        Returns one-time payment records for a tracking ID. Optionally filter by
        recipient and paginate with `limit` / `offset`.
      tags:
        - OneTime
      operationId: getOneTimePayment
      parameters:
        - in: path
          name: trackingId
          required: true
          schema:
            type: string
          description: Tracking ID assigned at checkout.
        - in: query
          name: recipient
          schema:
            type: string
            minLength: 32
            maxLength: 44
          description: Filter by recipient wallet public key.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 100
          description: Maximum records to return.
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of records to skip.
      responses:
        '200':
          description: Matching one-time payment record(s).
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                  - timestamp
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    oneOf:
                      - type: object
                        description: Single record (when exactly one matches).
                      - type: array
                        items:
                          type: object
                        description: Multiple records.
                  timestamp:
                    type: integer
                    description: Unix epoch ms.
        '400':
          description: Missing `trackingId`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No matching payment found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...